created Order Model and moved seed file under script
